refactor: isolate sidebar daemon
